INTERLOK-3561 Create superclass for PoolingWorkflow
Create a WorkflowWithObjectPool that provides some
abstracted behaviour for PoolingWorkflow.
No-Regression failures.
